Make eva-pipeline use classes of variation-commons

+++ src/test/java/uk/ac/ebi/eva/pipeline/model/EnsemblVariantTest.java
@@ -18,7 +18,7 @@ package uk.ac.ebi.eva.pipeline.model;
 
 import org.junit.Test;
 
-import uk.ac.ebi.eva.commons.models.data.Variant;
+import uk.ac.ebi.eva.commons.core.models.pipeline.Variant;
 import uk.ac.ebi.eva.pipeline.io.mappers.VariantVcfFactory;
 
 import java.util.List;
@@ -79,7 +79,7 @@ public class EnsemblVariantTest {
         List<Variant> result = factory.create(FILE_ID, STUDY_ID, line);
         assertEquals(1, result.size());
 
-        int start = result.get(0).getStart();
+        int start = (int) result.get(0).getStart();
         String reference = result.get(0).getReference();
         String alternate = result.get(0).getAlternate();
 
@@ -87,7 +87,7 @@ public class EnsemblVariantTest {
         assertEquals("", reference);
         assertEquals("A", alternate);
 
-        EnsemblVariant insertion = new EnsemblVariant(result.get(0).getChromosome(), start, result.get(0).getEnd(),
+        EnsemblVariant insertion = new EnsemblVariant(result.get(0).getChromosome(), start, (int) result.get(0).getEnd(),
                                                       reference, alternate);
         assertEquals(4, insertion.getStart());
         assertEquals(3, insertion.getEnd());
@@ -100,7 +100,7 @@ public class EnsemblVariantTest {
         List<Variant> result = factory.create(FILE_ID, STUDY_ID, line);
         assertEquals(1, result.size());
 
-        int start = result.get(0).getStart();
+        int start = (int) result.get(0).getStart();
         String reference = result.get(0).getReference();
         String alternate = result.get(0).getAlternate();
 
@@ -108,7 +108,7 @@ public class EnsemblVariantTest {
         assertEquals("C", reference);
         assertEquals("", alternate);
 
-        EnsemblVariant insertion = new EnsemblVariant(result.get(0).getChromosome(), start, result.get(0).getEnd(),
+        EnsemblVariant insertion = new EnsemblVariant(result.get(0).getChromosome(), start, (int) result.get(0).getEnd(),
                                                       reference, alternate);
         assertEquals(3, insertion.getStart());
         assertEquals(3, insertion.getEnd());
